split SquidGame from commandline code
